Add file locking since we now modify mbox.
This implementation is a little pessimistic (grabs a lock even if
no changes are to be made), but in this case the lock should not
be held long enough to matter.

I plan to add support for flock as well, because it seems that other
MUA's (like mailx on OpenBSD) use this instead of fcntl locking.

Unfortunately fcntl locks and flock seemingly cannot be used on the
same file, so one will have to take precedence over the other.