Fix a bug in cache handling.
When a cache entry is removed from the cache, its date field is set
to -1 to indicate that it can no longer be taken. Because cache_take
returned a pointer to this letter, the caller would receive a letter
with a date of -1, instead of the actual date of the letter. To avoid
this, copy the letter to an out parameter instead of returning a pointer
to it.

Additionally, do not read the letter from the disk after sucessfully
finding a cached letter, leading to 2 duplicate letters being stored.