Use the From address as a sentinel value instead of the date
Technically the on-disk cache could have a date with the value -1,
though the program currently would not generate such a date.

This avoids a memory leak if the on-disk cache has a date with the 
value -1, because mailz no longer erroniously believes that such a
letter has already been 'taken'.

The value of from->addr cannot be NULL, since it is a mandatory field