Ignore whitespace-only tokens in command input.
This is the typical paradigm used in shell-like programs, and it
helps in cases where a space was accidentally entered but the input
is otherwise valid.

This command input handling code is poor and probably requires a
rewrite as well as testing code written for it.

However, I dont think that this feature adds enough complexity that
it makes the code significantly worse, so for now it may as well go
in.