Treat snprintf(3) < 0 return the same as a >= sizeof(buf) return
The only way snprintf(3) can really fail is EOVERFLOW if the return
value would exceed INT_MAX.
This is effectively equivalent to the buffer being too small,
because the given size of the buffer cannot exceed INT_MAX either.
In the case where the return value is >= sizeof(buf) the return value
is not further inspected to determine the size of buffer that would be
required, so these cases can be treated equivalently.